Merging Dev

+++ src/qrl/core/config.py
@@ -56,7 +56,7 @@ class UserConfig(object):
         self.chain_state_broadcast_period = 30
         # must be less than ping_timeout
 
-        self.transaction_minimum_fee = 0 * DevConfig(ignore_check).shor_per_quanta
+        self.transaction_minimum_fee = int(0 * DevConfig(ignore_check).shor_per_quanta)
         self.transaction_pool_size = 25000
         self.pending_transaction_pool_size = 75000
         # 1% of the pending_transaction_pool will be reserved for moving stale txn
